# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Boxing Center — API réservation coachs
  version: 1.0.0
  description: |
    Contrat v1 (Europe/Paris). Lots : Eddy (créneaux/BO), Brad (UI), Raphael (paiement/signature/QR/Deciplus).
    Cahier lisible : CAHIER-API.md
  contact:
    name: Boxing Center
    email: boxingcenter31@gmail.com
servers:
  - url: https://coach.boxingcenter.fr/api/v1
    description: Production
  - url: http://localhost:3041/api/v1
    description: Local
...
security:
  - cookieAuth: []
paths:
...
  /reservations/{id}:
    get:
      tags: [reservations]
      operationId: getReservation
      parameters:
        - $ref: "#/components/parameters/ReservationId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /reservations/{id}/checkout:
    post:
      tags: [payments]
      summary: Checkout 1× Payplug / PayPal / avoir (Raphael)
      operationId: checkoutReservation
      parameters:
        - $ref: "#/components/parameters/ReservationId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider]
              properties:
                provider:
                  type: string
                  enum: [payplug, paypal, credit]
      responses:
        "200":
          description: URL checkout ou passage avoir
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "409":
          $ref: "#/components/responses/Error"
  /reservations/{id}/payment/sync:
    post:
      tags: [payments]
      summary: Re-query prestataire au retour navigateur
      operationId: syncPayment
      parameters:
        - $ref: "#/components/parameters/ReservationId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
  /webhooks/payplug:
    post:
      tags: [payments]
      security: []
      summary: Webhook Payplug (signature obligatoire)
      operationId: webhookPayplug
      parameters:
        - name: PayPlug-Signature
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        "200": { description: ACK }
        "401":
          $ref: "#/components/responses/Error"
  /webhooks/paypal:
    post:
      tags: [payments]
      security: []
      summary: Webhook PayPal
      operationId: webhookPaypal
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        "200": { description: ACK }
        "401":
          $ref: "#/components/responses/Error"
  /reservations/{id}/documents:
    get:
      tags: [signatures]
      summary: Docs à signer (URLs TTL)
      operationId: listDocuments
      parameters:
        - $ref: "#/components/parameters/ReservationId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  documents:
                    type: array
                    items:
                      $ref: "#/components/schemas/DocumentVersion"
        "409":
          $ref: "#/components/responses/Error"
  /reservations/{id}/signature:
    post:
      tags: [signatures]
      summary: Signer CGV + RI + décharge (Raphael)
      operationId: postSignature
      parameters:
        - $ref: "#/components/parameters/ReservationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignatureRequest"
      responses:
        "200":
          description: Réservation confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "409":
          $ref: "#/components/responses/Error"
...
  /reservations/{id}/qr:
    get:
      tags: [qr]
      summary: PNG QR opaque si confirmed
      operationId: getQr
      parameters:
        - $ref: "#/components/parameters/ReservationId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrPayload"
        "409":
          $ref: "#/components/responses/Error"
...
  /internal/access/verify:
    post:
      tags: [internal, qr]
      security:
        - syncSecret: []
      summary: Vérifier un token QR (service)
      operationId: verifyAccess
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, club_id]
              properties:
                token: { type: string }
                club_id:
                  $ref: "#/components/schemas/ClubId"
      responses:
        "200":
          description: Accès OK
          content:
            application/json:
              schema:
                type: object
                required: [ok, reservation_id, coach_id]
                properties:
                  ok: { type: boolean }
                  reservation_id: { type: string, format: uuid }
                  coach_id: { type: string, format: uuid }
        "401":
          $ref: "#/components/responses/Error"
        "410":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /internal/deciplus/jobs:
    post:
      tags: [internal]
      security:
        - syncSecret: []
      operationId: enqueueDeciplusJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action, reservation_id]
              properties:
                action:
                  type: string
                  enum: [grant, revoke, revoke_coach]
                reservation_id: { type: string, format: uuid }
                coach_id: { type: string, format: uuid }
      responses:
        "202": { description: En file }
  /internal/deciplus/callback:
    post:
      tags: [internal]
      security:
        - syncSecret: []
      operationId: deciplusCallback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reservation_id, job_status]
              properties:
                reservation_id: { type: string, format: uuid }
                deciplus_member_id: { type: string }
                job_status:
                  type: string
                  enum: [queued, granted, revoked, error]
                error: { type: string }
      responses:
        "200": { description: ACK }
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Cookie session Supabase httpOnly
    syncSecret:
      type: apiKey
      in: header
      name: x-sync-secret
  parameters:
...
    ReservationId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, format: uuid }
  responses:
    NotFound:
      description: Absent ou hors périmètre (anti-énumération)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Error:
      description: Erreur métier
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
  schemas:
    ClubId:
      type: string
      enum: [minimes, st-cyprien, etats-unis, ramonville, portet]
    ReservationStatus:
      type: string
      enum:
        - held
        - awaiting_signature
        - confirmed
        - consumed
        - expired
        - payment_failed
        - cancelled_credit
        - no_show
    PaymentStatus:
      type: string
      enum: [unpaid, paid, failed, waived_credit]
    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - UNAUTHENTICATED
        - FORBIDDEN
        - SUSPENDED
        - NOT_FOUND
        - SLOT_FULL
        - SLOT_BLOCKED
        - ACTIVE_LIMIT
        - HOLD_EXPIRED
        - PAYMENT_REQUIRED
        - SIGNATURE_REQUIRED
        - CANCEL_TOO_LATE
        - PRICE_MISMATCH
        - CONFLICT
        - QR_WINDOW_CLOSED
        - QR_WRONG_CLUB
        - RATE_LIMITED
        - WEBHOOK_INVALID
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              $ref: "#/components/schemas/ErrorCode"
            message: { type: string }
            details: { type: object, additionalProperties: true }
...
    Reservation:
      type: object
      required:
        - id
        - coach_id
        - club_id
        - space_id
        - starts_at
        - ends_at
        - amount_cents
        - status
        - payment_status
        - signature_status
      properties:
        id: { type: string, format: uuid }
        coach_id: { type: string, format: uuid }
        club_id:
          $ref: "#/components/schemas/ClubId"
        space_id: { type: string }
        starts_at: { type: string, format: date-time }
        ends_at: { type: string, format: date-time }
        amount_cents: { type: integer }
        currency: { type: string, default: eur }
        status:
          $ref: "#/components/schemas/ReservationStatus"
        payment_status:
          $ref: "#/components/schemas/PaymentStatus"
        payment_provider: { type: string, nullable: true }
        signature_status: { type: string, enum: [none, signed] }
        signed_at: { type: string, format: date-time, nullable: true }
        hold_expires_at: { type: string, format: date-time, nullable: true }
        qr_valid_from: { type: string, format: date-time, nullable: true }
        qr_valid_to: { type: string, format: date-time, nullable: true }
        qr_ready: { type: boolean }
        deciplus_job_status:
          type: string
          enum: [none, queued, granted, revoked, error]
        credit_id: { type: string, format: uuid, nullable: true }
        created_at: { type: string, format: date-time }
...
    CheckoutResponse:
      type: object
      properties:
        reservation_id: { type: string, format: uuid }
        provider: { type: string, enum: [payplug, paypal, credit] }
        checkout_url: { type: string, format: uri }
        status:
          $ref: "#/components/schemas/ReservationStatus"
    DocumentVersion:
      type: object
      required: [id, kind, title, version]
      properties:
        id: { type: string, format: uuid }
        kind: { type: string, enum: [cgv, reglement, decharge, other] }
        title: { type: string }
        version: { type: string }
        url: { type: string, format: uri }
    SignatureRequest:
      type: object
      required: [document_ids, signature_image, consent]
      properties:
        document_ids:
          type: array
          minItems: 3
          items: { type: string, format: uuid }
        signature_image:
          type: string
          description: data:image/png;base64,…
        consent: { type: boolean }
    QrPayload:
      type: object
      required: [png_data_url, valid_from, valid_to, club_id, state]
      properties:
        png_data_url: { type: string }
        valid_from: { type: string, format: date-time }
        valid_to: { type: string, format: date-time }
        club_id:
          $ref: "#/components/schemas/ClubId"
        state: { type: string, enum: [waiting, active, expired, revoked] }
...